feat(lexparse): parse and validate
lexparse now returns both a parsed spec and errors.
