+++ src/plccng/lexparse/__init__.py
@@ -1,10 +1,10 @@
-from .lexvalidate.check_format_of_patterns import InvalidPattern
-from .lexvalidate.check_format_of_names import InvalidName
-from .lexvalidate.check_for_duplicate_names import DuplicateName
 from .LexicalSpec import LexicalSpec
 from .LexicalRule import LexicalRule
-from .lexvalidate.check_for_unrecognized_lines import (
-    UnrecognizedLine,
+from .lexvalidate import (
+    InvalidName,
+    InvalidPattern,
+    DuplicateName,
+    UnrecognizedLine
 )
-from .lexparse import fromstring
-from .lexparse import fromlines
+from .lexparse import from_string
+from .lexparse import from_lines
