+++ src/plccng/lexparse/lexvalidate/check_format_of_names_test.py
@@ -3,7 +3,7 @@ from .check_format_of_names import check_format_of_names, InvalidName
 
 
 def test_all_upper_is_valid():
-    spec = lexparse.fromstring("""
+    spec = lexparse._from_string_without_validation("""
         token GOOD 'hi'
     """)
     e = check_format_of_names(spec.ruleList)
@@ -11,7 +11,7 @@ def test_all_upper_is_valid():
 
 
 def test_underscores_are_valid():
-    spec = lexparse.fromstring("""
+    spec = lexparse._from_string_without_validation("""
         token _GOOD_NAME_ 'hi'
     """)
     e = check_format_of_names(spec.ruleList)
@@ -19,7 +19,7 @@ def test_underscores_are_valid():
 
 
 def test_numbers_are_valid():
-    spec = lexparse.fromstring("""
+    spec = lexparse._from_string_without_validation("""
         token GOOD_1 'hi'
     """)
     e = check_format_of_names(spec.ruleList)
@@ -27,7 +27,7 @@ def test_numbers_are_valid():
 
 
 def test_lowercase_are_invalid():
-    spec = lexparse.fromstring("""
+    spec = lexparse._from_string_without_validation("""
         token bad_name 'hi'
     """)
     e = check_format_of_names(spec.ruleList)
@@ -35,7 +35,7 @@ def test_lowercase_are_invalid():
 
 
 def test_starting_with_a_number_is_invalid():
-    spec = lexparse.fromstring("""
+    spec = lexparse._from_string_without_validation("""
         token 1_BAD_NAME 'hi'
     """)
     e = check_format_of_names(spec.ruleList)
@@ -43,7 +43,7 @@ def test_starting_with_a_number_is_invalid():
 
 
 def test_hyphens_are_invalid():
-    spec = lexparse.fromstring("""
+    spec = lexparse._from_string_without_validation("""
         token BAD-NAME 'hi'
     """)
     e = check_format_of_names(spec.ruleList)
@@ -51,7 +51,7 @@ def test_hyphens_are_invalid():
 
 
 def test_any_other_punctuation_is_invalid():
-    spec = lexparse.fromstring("""
+    spec = lexparse._from_string_without_validation("""
         token BAD!NAME 'hi'
     """)
     e = check_format_of_names(spec.ruleList)
